REPORT ON COOKING USING FIREWOOD
Cooking with firewood in form of logs and branches from trees is the oldest cooking fuel. It serves purposes such as cooking, heating. Firewood can be categorized into hardwood and softwood; Softwoods burn more quickly and generate less heat. The drier the firewood, the less energy is available for heating or cooking purposes.